+++ src/arraydialog.h
@@ -25,10 +25,31 @@ public:
 	Ui::ArrayDialog ui;
 	QString getLatexText();
 
+private:
+	QStringList packageList = {"latex", "amsmath", "mathtools"};
+	QStringList environmentList = {"array", "alignedat", "matrix", "pmatrix", "bmatrix", "Bmatrix", "vmatrix", "Vmatrix"};
+	QList<Qt::AlignmentFlag> alignList = {Qt::AlignHCenter, Qt::AlignLeft, Qt::AlignRight};
+	QStringList alignTextList = {tr("Center"), tr("Left"), tr("Right")};
+	QList<int> arrayAligns;
+	void addEmptyTableItems();
+	void addEmptyColumnItems(int col);
+	void setComboEnv(QString package);
+	void setColAlignments();
+	void setColAlignment(int col, Qt::AlignmentFlag align);
+	int currentColAlignIndex(int col);
+	QStringList getEnvBeginEndStatements();
+	QString getFactoredAlignments();
+	QString squash(QString part, QString align);
+
 protected slots:
 	void newRows(int num);
 	void newColumns(int num);
     void keyPressEvent(QKeyEvent *event);
+	void setTitle();
+	void slotEnvironmentChanged();
+	void slotCurrentCellChanged(int row, int col);
+	void slotAlignmentChanged(int ix);
+	void slotPackageChanged();
 };
 
 #endif